requirements.txt
apache-beam==2.60.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.9.7
fasteners==0.19
Flask==3.0.3
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
numpy==1.26.4
objsize==0.7.0
orjson==3.10.11
packaging==24.2
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.10
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
zstandard==0.23.0
